move decorators to structures interface
