+++ autoarray/__init__.py
@@ -81,7 +81,7 @@ from .structures.mesh.delaunay_2d import Mesh2DDelaunay
 from .structures.arrays.kernel_2d import Kernel2D
 from .structures.vectors.uniform import VectorYX2D
 from .structures.vectors.irregular import VectorYX2DIrregular
-from .structures import structure_decorators as grid_dec
+from .structures.interface import decorators as grid_dec
 from .structures.header import Header
 from .layout.region import Region1D
 from .layout.region import Region2D
