+++ test_autoarray/operators/over_sample/test_iterate.py
@@ -2,7 +2,7 @@ import numpy as np
 
 import autoarray as aa
 
-from autoarray.structures.mock.mock_structure_decorators import (
+from autoarray.structures.mock.mock_decorators import (
     ndarray_1d_from,
     ndarray_1d_zeros_from,
     ndarray_2d_from,
